+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Added support for SQLite. #51
   - Added support for Oracle Database. #13
   - Added History Explorer on sidebar
+  - Added Bookmarks Explorer on sidebar
   - MySQL Xdevapi: Added support to MySQL XDEVAPI. It fixes `ER_NOT_SUPPORTED_AUTH_MODE: Client does not support authentication protocol` error. (Experimental).
   - Export query results as JSON and CSV. #95
     - @TODO: Add image
@@ -27,6 +28,7 @@
     - `SQLTools.appendToCursor` command was renamed to `SQLTools.insertText`
     - `SQLTools.refreshSidebar` command was renamed to `SQLTools.refreshAll`
     - `SQLTools.runFromInput` command was renamed to `SQLTools.executeFromInput`
+    - `SQLTools.editFromHistory` command was renamed to `SQLTools.editHistory`
   - ***Settngs***:
     - `sqltools.logging` and `sqltools.logLevel` were removed.